[ci:force] move lock and favorite
