+++ docs/upgrade-guide/upgrade50-60.md
@@ -126,6 +126,8 @@ v6.0.0 and after:
 | [`NodeFavoriteDirective`](../content-services/directives/node-favorite.directive.md)| `@alfresco/adf-core` | `@alfresco/adf-content-services` |
 | [`NodeRestoreDirective`](../content-services/directives/node-restore.directive.md)| `@alfresco/adf-core` | `@alfresco/adf-content-services` |
 | [`AppsProcessService`] | `@alfresco/adf-core` | `@alfresco/adf-process-services` |
+| `LockService` | `@alfresco/adf-core` | `@alfresco/adf-process-services` |
+| `FavoritesApiService` | `@alfresco/adf-core` | `@alfresco/adf-process-services` |
 
 ### Update Data-table a data change
