+++ lib/content-services/src/lib/common/services/favorites-api.service.ts
@@ -18,8 +18,7 @@
 import { Injectable } from '@angular/core';
 import { FavoritesApi, NodePaging, FavoritePaging } from '@alfresco/js-api';
 import { Observable, from, of } from 'rxjs';
-import { AlfrescoApiService } from './alfresco-api.service';
-import { UserPreferencesService } from '../common/services/user-preferences.service';
+import { AlfrescoApiService, UserPreferencesService } from '@alfresco/adf-core';
 import { catchError } from 'rxjs/operators';
 
 @Injectable({
