+++ lib/content-services/src/lib/content-node-selector/content-node-selector-panel.component.spec.ts
@@ -52,7 +52,7 @@ import { TranslateModule } from '@ngx-translate/core';
 import { SearchQueryBuilderService } from '../search';
 import { ContentNodeSelectorPanelService } from './content-node-selector-panel.service';
 import { mockContentModelTextProperty } from '../mock/content-model.mock';
-import { SitesService } from '../services/sites.service';
+import { SitesService } from '../common/services/sites.service';
 
 const fakeResultSetPaging: ResultSetPaging = {
     list: {
