+++ lib/content-services/src/lib/document-list/public-api.ts
@@ -38,6 +38,7 @@ export * from './services/document-actions.service';
 export * from './services/document-list.service';
 export * from './services/node-actions.service';
 export * from './services/custom-resources.service';
+export * from './services/lock.service';
 
 // models
 export * from './models/content-action.model';
