+++ lib/content-services/src/public-api.ts
@@ -40,6 +40,6 @@ export * from './lib/new-version-uploader';
 export * from './lib/interfaces/index';
 export * from './lib/version-compatibility/index';
 export * from './lib/pipes/index';
-export * from './lib/services/index';
+export * from './lib/common/index';
 
 export * from './lib/content.module';
