+++ lib/core/src/lib/services/public-api.ts
@@ -26,7 +26,6 @@ export * from './upload.service';
 export * from './dynamic-component-mapper.service';
 export * from './highlight-transform.service';
 export * from './deleted-nodes-api.service';
-export * from './favorites-api.service';
 export * from './nodes-api.service';
 export * from './people-content.service';
 export * from './people-process.service';
@@ -37,7 +36,6 @@ export * from './comment-content.service';
 export * from './login-dialog.service';
 export * from './external-alfresco-api.service';
 export * from './download-zip.service';
-export * from './lock.service';
 export * from './automation.service';
 export * from './automation.service';
 export * from './download.service';
